backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
gunicorn==21.2.0
pytest==7.4.0
supabase==2.0.0
python-jose==3.3.0
cryptography==41.0.3
